Using a global ES docker container that is only created once durign the execution of oak-search-elastic
