+++ oak-search-elastic/src/test/java/org/apache/jackrabbit/oak/plugins/index/elastic/ElasticConnectionRule.java
@@ -17,6 +17,7 @@
 package org.apache.jackrabbit.oak.plugins.index.elastic;
 
 import com.github.dockerjava.api.DockerClient;
+import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
@@ -52,14 +53,14 @@ public class ElasticConnectionRule extends ExternalResource {
 
     private static final Logger LOG = LoggerFactory.getLogger(ElasticConnectionRule.class);
 
-    private static final String INDEX_PREFIX = "elastic_test";
-    private static final String PLUGIN_DIGEST = "db479aeee452b2a0f6e3c619ecdf27ca5853e54e7bc787e5c56a49899c249240";
+    private final String indexPrefix;
     private static boolean useDocker = false;
 
     private final String elasticConnectionString;
 
     public ElasticConnectionRule(String elasticConnectionString) {
         this.elasticConnectionString = elasticConnectionString;
+        indexPrefix = "elastic_test_" + RandomStringUtils.random(5, true, false).toLowerCase();
     }
 
     public ElasticsearchContainer elastic;
@@ -70,64 +71,11 @@ public class ElasticConnectionRule extends ExternalResource {
     @Override
     public Statement apply(Statement base, Description description) {
         Statement s = super.apply(base, description);
-        // see if docker is to be used or not... initialize docker rule only if that's the case.
-        final String pluginVersion = "7.16.3.0";
-        final String pluginFileName = "elastiknn-" + pluginVersion + ".zip";
-        final String localPluginPath = "target/" + pluginFileName;
-        downloadSimilaritySearchPluginIfNotExists(localPluginPath, pluginVersion);
         if (elasticConnectionString == null || getElasticConnectionFromString() == null) {
-            checkIfDockerClientAvailable();
-            Network network = Network.newNetwork();
-
-            elastic = new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:" + Version.CURRENT)
-                    .withCopyFileToContainer(MountableFile.forHostPath(localPluginPath), "/tmp/plugins/" + pluginFileName)
-                    .withCopyFileToContainer(MountableFile.forClasspathResource("elasticstartscript.sh"), "/tmp/elasticstartscript.sh")
-                    .withCommand("bash /tmp/elasticstartscript.sh")
-                    .withNetwork(network);
-            elastic.start();
-
+            elastic = ElasticTestServer.getESTestServer().container();
             setUseDocker(true);
         }
         return s;
-    }
-
-    @Override
-    protected void after() {
-        if (elastic != null && elastic.isRunning()) {
-            elastic.stop();
-        }
-    }
-
-    private void downloadSimilaritySearchPluginIfNotExists(String localPluginPath, String pluginVersion) {
-        File pluginFile = new File(localPluginPath);
-        if (!pluginFile.exists()) {
-            LOG.info("Plugin file {} doesn't exist. Trying to download.", localPluginPath);
-            try (CloseableHttpClient client = HttpClients.createDefault()) {
-                HttpGet get = new HttpGet("https://github.com/alexklibisz/elastiknn/releases/download/" + pluginVersion
-                        + "/elastiknn-" + pluginVersion + ".zip");
-                CloseableHttpResponse response = client.execute(get);
-                InputStream inputStream = response.getEntity().getContent();
-                MessageDigest messageDigest = MessageDigest.getInstance("SHA-256");
-                DigestInputStream dis = new DigestInputStream(inputStream, messageDigest);
-                FileOutputStream outputStream = new FileOutputStream(pluginFile);
-                IOUtils.copy(dis, outputStream);
-                messageDigest = dis.getMessageDigest();
-                // bytes to hex
-                StringBuilder result = new StringBuilder();
-                for (byte b : messageDigest.digest()) {
-                    result.append(String.format("%02x", b));
-                }
-                if (!PLUGIN_DIGEST.equals(result.toString())) {
-                    String deleteString = "Downloaded plugin file deleted.";
-                    if (!pluginFile.delete()) {
-                        deleteString = "Could not delete downloaded plugin file.";
-                    }
-                    throw new RuntimeException("Plugin digest unequal. Found " + result + ". Expected " + PLUGIN_DIGEST + ". " + deleteString);
-                }
-            } catch (IOException | NoSuchAlgorithmException e) {
-                throw new RuntimeException("Could not download similarity search plugin", e);
-            }
-        }
     }
 
     public ElasticConnection getElasticConnectionFromString() {
@@ -137,7 +85,7 @@ public class ElasticConnectionRule extends ExternalResource {
             String scheme = uri.getScheme();
             int port = uri.getPort();
             String query = uri.getQuery();
-
+            
             String api_key = null;
             String api_secret = null;
             if (query != null) {
@@ -145,7 +93,7 @@ public class ElasticConnectionRule extends ExternalResource {
                 api_secret = query.split(",")[1].split("=")[1];
             }
             return ElasticConnection.newBuilder()
-                    .withIndexPrefix(INDEX_PREFIX + System.currentTimeMillis())
+                    .withIndexPrefix(indexPrefix + System.currentTimeMillis())
                     .withConnectionParameters(scheme, host, port)
                     .withApiKeys(api_key, api_secret)
                     .build();
@@ -161,22 +109,11 @@ public class ElasticConnectionRule extends ExternalResource {
 
     public ElasticConnection getElasticConnectionForDocker(String containerIpAddress, int port) {
         return ElasticConnection.newBuilder()
-                .withIndexPrefix(INDEX_PREFIX + System.currentTimeMillis())
+                .withIndexPrefix(indexPrefix + System.currentTimeMillis())
                 .withConnectionParameters(ElasticConnection.DEFAULT_SCHEME,
                         containerIpAddress, port)
                 .withApiKeys(null, null)
                 .build();
-    }
-
-    private void checkIfDockerClientAvailable() {
-        DockerClient client = null;
-        try {
-            client = DockerClientFactory.instance().client();
-        } catch (Exception e) {
-            LOG.warn("Docker is not available and elasticConnectionDetails sys prop not specified or incorrect" +
-                    ", Elastic tests will be skipped");
-        }
-        assumeNotNull(client);
     }
 
     private void setUseDocker(boolean useDocker) {
